タイムテーブルのスクレイピングをTimetableモデルで行う
